Add option to randomize arguments, fix unescape()
This adds two new options, -a and -u. The -a option allows one to
randomize the order of the arguments, and -u does the same but unescapes
them first.

While there, fix unescape(): there were some bugs, and it didn't
properly handle short escape sequences.

Examples:
$ { echo a; echo b; echo c; } | ./randomize
a
b
c
$ ./randomize -a a b c
b
c
a
$ ./randomize -u a '\n' c

a
c