Complete rewrite
This uses a more flexible API to access files. After quite a bit of
work, this is again minimally viable. It doesn't do everything the
original did, but already improves in one area: it can read from a file
without slurping the whole file into memory (it essentially only keeps
pointers in memory).