Move from groff to mandoc
OpenBSD now uses mandoc, so use it instead of groff (with an option to
use groff instead).

Update a construct that causes mandoc to choke (.Os \&); while there,
set the update date to Aug 11 (commit
414d6f9eda2ee294f0ba2f5306c0b942e2e9b824). This has been fixed (see
mid.gmane.org/20101027192859.GJ16057@iris.usta.de), but the version in
OpenBSD 4.8 will still have this bug so leave the fix in.