Make randomize -n1 largefile more efficient
Introduce a new and undocumented temporary flag -p to switch to a new
strategy: pick a random byte from a file, and get the next record if the
byte matches what we're looking for.

Basic tests pass, but this should work with more than one file (and
there are lots of XXX to solve). This should just automatically be used
when possible, but it's not obvious when it works, let alone when it
helps. (Note that even a simple pattern like 'aa' may require us to read
the entire file up to the selected point, if it consists entirely of 'a'
characters.)