Handle EOF properly
We could previously call pcre_exec() with the PCRE_NOTEOL flag if we did
actually have the (whole/end of the) file in memory. Fix that, by not
handing the last byte we know to pcre_exec() except on EOF.

It would take a really nasty regex to notice the difference, but let's
be correct.